Good: Everything now compiles and runs, with the added bonus of NOT creating an infinite loop that crashes Unity.
Bad: Pathfinding does not work or reach destination Tile. I understand that this is based on a flaw in my logic however and intentionally left it until I had the earlier issues I was experiencing solved. Not exactly sure what I need to do but it shouldn't be too hard to debug now that everything runs and doesn't crash