made some quick fixes during my check-in with Adam today and then attempted to remove all of my repeat code in favor of iterating through an array and a list. When I iterate through everything though, Unity seems to enter an infinite loop and crash. When I only iterate through position of adjacent tiles (instead of iterating through the positions to make a list of all adjacent tile objects nad then adding those to paths and checking them) my program doesn't crash, however it doesn't find the correct path either.
Not sure what is going wrong since the min heap should be pushing the lowest weight path to the top.

My main thought is that since I'm adding four paths to the min heap every pass of the while loop (until tile with end position is found), the first few paths with the lowest weight are being pushed to the top so the end tile is never actually reached. I think finding some way to check proximity to the end Tile and remove the three least good paths every pass would help to ensure this doesn't happen but I don't have any ideas on how to implement that.